patient-service/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.1
pydantic_core==2.33.0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
